added merchandise row to event model (True or false if there is merchandise) added merchandise row to rider event profile (textarea for listing what they ordered for this event)
May need to add multiple event merchandise rows to the event model, for other merch.(hats, sirts, hoodies, etc) so I can load these items from the DB for each instance of an event.

Currently this is a 'static' merchandise page, not a good thing!!< will be priority to fix.