Calculating the age on the day of the event:
It's calculated on 'create' and on watch (on change)
I need to get the birth_date in put from the user, but the way I've used django formset, I don't have direct access to the input through v-model.
I tried adding the attribute (attrs={'v-model':'birth_date'}, but this doesn't function. I think because this model is not created on instantiation of vuejs instance.

need to calculate the age on the day of the event(Active)
need to write a test to stress test the server/ database queue times for 100+ users (maybe via selenium?) (Active)
need to add items to the paypal checkout list(Started, but Paused)