Ran a script to change all rider profiles first and last name and email to lower case and changed all confirmation numbers to uppercase
This is to clean the data so it's more easily searchable

todo: Need to return a list of riders registered with that confirmation number.

todo: Build the back end to query the database given first and last name or confirmation number

------ Create Confirmation Checking Page todo -------------

todo: allow users to see if they are registered for a race via confirmation number, first and last name

------------- Event Registration Page todo ----------------

todo: add more detail to the paypal cart

todo: change the confirmation number to match the paypal confirmation number

todo: change the confirmation email to non-html email

todo: allow the discount of an entry fee, if more than one rider is registered and a discount code is used
or todo: remove the paypal button and reveal the submit button if there is only one rider.

-------------------- Merch Page todo ----------------------

todo: Remove from max available on order from database (should do it on current state via add and remove functions)

**currently using flat 8% for shipping

todo: add a cost estimator API (Async) or

todo: Store data in database then email purchaser

then

notify someone to fulfill order