added the promotion to the models, frontend and back end
------------- Event Registration Page todo ----------------
todo: add a feature to allow a discount code to be used and it removes that total amount from the total cost of the cart. Cost of the cart can not == $0.00

todo: add more detail to the paypal cart

todo: allow the discount of an entry fee, if more than one rider is registered and a discount code is used
or todo: remove the paypal button and reveal the submit button if there is only one rider.