changed promo model to include description. logic on registration (per form) added to display the option to accept the propmo function not being called
todo: Send the data to the form (registration page)

todo: Add the JS to the front form if amateur is selected to reveal the promo if

todo: Get the data from the front to the back and remove the quantity on form submit

------------- Event Registration Page todo ----------------
todo: add a feature to allow a discount code to be used and it removes that total amount from the total cost of the cart. Cost of the cart can not == $0.00

todo: add more detail to the paypal cart

todo: allow the discount of an entry fee, if more than one rider is registered and a discount code is used
or todo: remove the paypal button and reveal the submit button if there is only one rider.