added discount codes and promotion data to the return, if there are errors in form
todo: figure out how to allow times. Probably in a list from the admin then make choice selectors. will have to count down the times slots as well..

todo: Get the data from the front to the back and remove the quantity on form submit

------------- Event Registration Page todo ----------------
todo: add a feature to allow a discount code to be used and it removes that total amount from the total cost of the cart. Cost of the cart can not == $0.00

todo: add more detail to the paypal cart

todo: allow the discount of an entry fee, if more than one rider is registered and a discount code is used
or todo: remove the paypal button and reveal the submit button if there is only one rider.