Promotion limits set on the home page and the times are limited to an equal division of the avaliable options for a promotion.(i.e. 3 options - promo limit 3 = 1 rider per option) Also removed unused code and print statements
This means I do not reduce the available riders number for the todo below:
todo: Get the data from the front to the back and remove the quantity on form submit

I can use this similar approach to the code discount todo.

------------- Event Registration Page todo ----------------
todo: add a feature to allow a discount code to be used and it removes that total amount from the total cost of the cart. Cost of the cart can not == $0.00

todo: add more detail to the paypal cart

todo: allow the discount of an entry fee, if more than one rider is registered and a discount code is used
or todo: remove the paypal button and reveal the submit button if there is only one rider.