Discount code per forms, will now correctly apply and remove the checkout button or the submit button.
todo: logic for removing and adding forms regarding not being able to use a code more than once

todo: wire up the price change

todo: if there is only one person registering and the discount is 100% then paypal will not allow checkout. I can either:
use the code to get into a single registration form page (special page) or unlock the submit button and submit the form.

If I unlock the submit button, I must retain a list of used codes on the page, so the code can't be used on multiple forms, because the code data is sent over from the back end. This data will not be changed dynamically.

------------- Event Registration Page todo ----------------
todo: add a feature to allow a discount code to be used and it removes that total amount from the total cost of the cart. Cost of the cart can not == $0.00

todo: add more detail to the paypal cart

todo: allow the discount of an entry fee, if more than one rider is registered and a discount code is used
or todo: remove the paypal button and reveal the submit button if there is only one rider.