fixed bug that allowed success msg when a duplicate discount code want used but not valid. limited the range of discount valued to .05 and 1 allowed for the import of discount codes through xlsx and csv files removed two unused fields from the registration model
Complete: backend logic for the single form no-paypal needs to be setup

Complete: logic for removing and adding forms regarding not being able to use a code more than once

Complete: wire up the price change

Complete: if there is only one person registering and the discount is 100% then paypal will not allow checkout. I can either:
use the code to get into a single registration form page (special page) or unlock the submit button and submit the form.

Complete: If I unlock the submit button, I must retain a list of used codes on the page, so the code can't be used on multiple forms, because the code data is sent over from the back end. This data will not be changed dynamically.

------------- Event Registration Page todo ----------------
Complete: add a feature to allow a discount code to be used and it removes that total amount from the total cost of the cart. Cost of the cart can not == $0.00

todo: add more detail to the paypal cart

todo: allow the discount of an entry fee, if more than one rider is registered and a discount code is used
or todo: remove the paypal button and reveal the submit button if there is only one rider.